% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ageplot.R
\name{ageplot}
\alias{ageplot}
\title{Create a complete age profile ggplot appropriate to a particular data type}
\usage{
ageplot(object, ...)
}
\arguments{
\item{object}{an object, whose class will determine the behaviour of \code{ageplot}}

\item{...}{other arguments passed to specific methods}
}
\value{
a ggplot object
}
\description{
\code{ageplot} uses ggplot2 to draw a particular plot for an object of a
particular class in a single command. This defines the S3 generic that
other classes and packages can extend.
}
\seealso{
\code{\link{ggplot}} and \code{\link{fortify}}
}

