% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autocast.R
\docType{package}
\name{AutoCast}
\alias{AutoCast}
\alias{AutoCast-package}
\title{Automated Bayesian Forecasting with YourCast}
\description{
Generates Yourcasts under range of prior specifications, ranking forecasts by user preferences for smoothness and out-of-sample fit.
}

