% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Run YourCast forecasts given selected weights.}
\usage{
forecast(obj, weights, point.estimate = median, lower.bound = min,
  upper.bound = max)
}
\arguments{
\item{obj}{Object of class \code{"gridcast"}.}

\item{weights}{Object of class \code{"weightcast"}.}

\item{point.estimate}{Function applied to calculate the point estimate of the final forecast. Default: \code{median}.}

\item{lower.bound}{Function applied to calculate the lower bound the final forecast. Default: \code{min}.}

\item{upper.bound}{Function applied to calculate the upper bound of the final forecast. Default: \code{max}.}
}
\value{
\code{forecast} returns an object of class \code{"forecast"}, which contains the following components:
\itemize{
     \item \code{yhat} A matrix of forecast point estimates (age groups by time periods).
     \item \code{yhat.lower} A matrix of forecast lower bound estimates (age groups by time periods).
     \item \code{yhat.upper} A matrix of forecast upper bound estimates (age groups by time periods).
     \item \code{y} A matrix of observed data (age groups by time periods).
     \item \code{aux} List. A list of inputs to \code{yourcast}.
     \item \code{validation} List. A list containing all sigma combinations and the associated diagnostics.
     \item \code{weights} The \code{"weightcast"} object provided as an input to \code{forecast}.
  }
}
\description{
\code{forecast} runs final YourCast forecasts given a \code{"gridcast"} object and a \code{"weightcast"} object as inputs.
}

