% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gridcast.R
\name{gridcast}
\alias{gridcast}
\title{Automated Bayesian forecasting with YourCast}
\usage{
gridcast(dataobj, Ha.sigma.seq = c(from = 0.01, to = 20, length.out = 5),
  Ha.list = NULL, Ht.sigma.seq = c(from = 0.01, to = 20, length.out = 5),
  Ht.list = NULL, Hat.sigma.seq = NA, Hat.list = NULL, logscale = TRUE,
  time.degree = 1, length.holdout = 5, holdout = NULL,
  mse.age.weight = NA, print.runs = TRUE, ...)
}
\arguments{
\item{dataobj}{Object of class '\code{yourprep}'}

\item{Ha.sigma.seq}{Three element numeric vector. The first two elements give the range of \code{Ha.sigma} values to search over.
The 3rd element gives the number of values to sample from the range.
If do not want to use age smoothing, set as \code{NA} and leave \code{Ha.list} as \code{NULL}.}

\item{Ha.list}{Vector. A vector of additional values of \code{Ha.sigma} to include in the grid search.Default: \code{NULL}.}

\item{Ht.sigma.seq}{Three element numeric vector. The first two elements give the range of \code{Ht.sigma} values to search over.
The 3rd element gives the number of values to sample from the range.
If do not want to use time smoothing, set as \code{NA} and leave \code{Ht.list} as \code{NULL}.}

\item{Ht.list}{Vector. A vector of additional values of \code{Ht.sigma} to include in the grid search. Default: \code{NULL}.}

\item{Hat.sigma.seq}{Three element numeric vector. The first two elements give the range of \code{Hat.sigma} values to search over.
 The 3rd element gives the number of values to sample from the range.
If do not want to use trend smoothing, set as \code{NA} and leave \code{Hat.list} as \code{NULL}.}

\item{Hat.list}{Vector. A vector of additional values of \code{Hat.sigma} to include in the grid search. Default: \code{NULL}.}

\item{logscale}{Logical. Should sigma values be evenly sampled on a log scale and then exponentiated? Default: \code{TRUE}.}

\item{time.degree}{Non-negative integer. Specifies the degree of the baseline polynomial to which time profiles are smoothed.
For example, if \code{time.degree=1}, then the forecasts cllosest to a straight line are scored highest.
If \code{time.degree}=0, then forecasts closest to a flat line are scored highest.}

\item{length.holdout}{Non-negative integer. Specifies the number of time periods that should be omitted from the end of the observed period for validation.
If set to 0, then at least one time period must be provided as an argument to \code{holdout}.}

\item{holdout}{Vector specifying additional time periods to withhold for validation. Default: \code{NULL}.}

\item{mse.age.weight}{A scalar or a numeric vector with weights that determine how much the different age groups factor into the calculation of root mean square error of the forecasts in the validation period.
If set to \code{0} or \code{NA}, age groups are weighted equally;
if set to a nonzero scalar, the weight for age group a is set proportional to a^\code{mse.age.weight};
if a vector of length A, the ath element is the weight of age group a. Default: \code{0}.}

\item{print.runs}{Logical. If TRUE, will print notification for each run of yourcast.}

\item{...}{Additional arguments passed to \code{yourcast} function.}
}
\value{
\code{panelAR} returns an object of class \code{"gridcast"}.

An object of class \code{"gridcast"} contains the following components
\itemize{
     \item \code{y} A matrix of observed data (age groups by time periods).
     \item \code{aux} List. A list of inputs to \code{yourcast}.
     \item \code{validation} List. A list containing all sigma combinations, the forecasts, and the associated diagnostics.
  }
}
\description{
\code{gridcast} generates YourCast forecasts for a range of prior specifications and outputs diagnostics across the different specifications.
}

