% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autocast.R
\name{netherlands_data}
\alias{netherlands_data}
\title{Deaths from breast cancer in the Netherlands}
\format{A \code{yourprep} object with 12 cross sections (age groups)}
\source{
World Health Organization. Available from Federico Girosi and Gary King, 2007, "Cause of Death Data", \url{http://hdl.handle.net/1902.1/UOVMCPSWOL}.
}
\description{
A \code{yourprep} object containing breast cancer deaths across 12 age groups in the Netherlands for 1950--2000.
Data contains covariates that are lagged by 30 years. The variables in the data are as follows:
}
\details{
\itemize{
  \item brst3. Breast cancer deaths.
  \item popu3. Population.
  \item tobacco2. Tobacco consumption.
  \item gdp. GDP.
  \item fat. Fat consumption.
  \item hc. Measure of human capital.
  \item time. Year covariates measured.
  ...
}
}
\references{
Girosi, Federico, and Gary King. 2008. Demographic Forecasting. Princeton: Princeton University Press. Copy at \url{http://j.mp/pqms4U}.
}

