% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autocast.R
\name{netherlands_weights}
\alias{netherlands_weights}
\title{Selected optimal weights for forecast of breast cancer in the Netherlands}
\format{A \code{weightcast} object with a single selected weight combination.}
\description{
A \code{weightcast} object containing a selected weight combination for the forecast of breast cancer in the Netherlands. See \code{\link{weightcast}} for additional details.
}

