% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.forecast.R
\name{plot.forecast}
\alias{plot.forecast}
\title{Plot AutoCast forecasts.}
\usage{
\method{plot}{forecast}(x, type = "agetime", title = NULL,
  subtitle = NULL, ageprofile.opts = list(), timeprofile.opts = list(),
  print = "device", print.opts = list())
}
\arguments{
\item{x}{Object of class \code{"forecast"}.}

\item{type}{Character. Type of plot: "age", "time", or "agetime". Default: "agetime".}

\item{title}{Character. Title to be used for plot. Default: \code{NULL}.}

\item{subtitle}{Character. Subtitle to be used for plot. Default: \code{NULL}.}

\item{ageprofile.opts}{List.}

\item{timeprofile.opts}{List.}

\item{print}{Character. Must be either "device" or "pdf". Default: "device".}

\item{print.opts}{List.}
}
\description{
Create plots of age and time profiles of forecasts.
}

