% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timeplot.R
\name{timeplot}
\alias{timeplot}
\title{Create a complete time profile ggplot appropriate to a particular data type}
\usage{
timeplot(object, ...)
}
\arguments{
\item{object}{an object, whose class will determine the behaviour of \code{timeplot}}

\item{...}{other arguments passed to specific methods}
}
\value{
a ggplot object
}
\description{
\code{timeplot} uses ggplot2 to draw a particular plot for an object of a
particular class in a single command. This defines the S3 generic that
other classes and packages can extend.
}
\seealso{
\code{\link{ggplot}} and \code{\link{fortify}}
}

