% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weightcast.R
\name{weightcast}
\alias{weightcast}
\title{Weight selection for AutoCast}
\usage{
weightcast(x, out.file = NULL, out.object = NULL, center.diags = T,
  scale.diags = T, options = list())
}
\arguments{
\item{x}{Object of class \code{"gridcast"} or a list of \code{"gridcast"} objects.}

\item{out.file}{Character. Gives full filename (and path) of where to save selected weight combinations (if requested). Default: NULL.}

\item{out.object}{Character. Gives name of object in \code{parent.frame()} to which to assign selected weight combinations (if requested). Default: NULL.}

\item{center.diags}{Logical. If \code{TRUE}, each of the four diagnostic variables is centered (de-meaned). Default: \code{TRUE}.}

\item{scale.diags}{Logical. If \code{TRUE}, each of the four diagnostic variables is scaled after centering, if applicable (divided by standard deviation).
Default: \code{TRUE}.}

\item{options}{Options passed to \code{options} argument of \code{ShinyApp}.}
}
\value{
\code{NULL}.
}
\description{
\code{weightcast} launches an interactive, Shiny-based application that allows the user to select multiple weight combinations that fit her preference for fit versus smoothness of the forecasts.
}

